[Update](inverted index) set default inverted index storage format to v2 (#35303)
This pull request changes the default storage format for inverted
indexes to V2. The V2 format enhances efficiency by merging all index
files within a single segment into one file. This update aims to reduce
IO cost when writing inverted index files, especially for cloud mode.

+++ be/src/olap/rowset/segment_v2/inverted_index_file_reader.cpp
@@ -73,7 +73,7 @@ Status InvertedIndexFileReader::_init_from_v2(int32_t read_buffer_size) {
             ReaderFileEntry* entry = nullptr;
 
             for (int32_t i = 0; i < numIndices; ++i) {
-                int64_t indexId = _stream->readInt();       // Read index ID
+                int64_t indexId = _stream->readLong();      // Read index ID
                 int32_t suffix_length = _stream->readInt(); // Read suffix length
                 std::vector<uint8_t> suffix_data(suffix_length);
                 _stream->readBytes(suffix_data.data(), suffix_length);
